+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# v1.1.0 ( 10 dec 2017 )
+Export to array form and import from it
+* AssociativeArray -> Array form -> AssociativeArray
+
 # v1.0.0 ( 9 dec 2017 )
 Hello world!